requirements.txt
pandas==2.0.3
tqdm==4.66.1
tensorboard==2.14.0
torch==2.1.2
torchvision==0.16.2
hydra-core==1.3.2
omegaconf==2.3.0
scikit-learn==1.3.2